pyproject.toml
[tool.poetry]
name = "live-ad-detection"
version = "0.1.0"
description = "Live advertisement detection cluster with touchscreen setup interface"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
flask = "^3.0.0"
qrcode = {extras = ["pil"], version = "^7.4.2"}
kivy = "^2.3.0"
python-wifi = "^0.6.1"
netifaces = "^0.11.0"
python-networkmanager = "^2.2"
pillow = "^10.1.0"
requests = "^2.31.0"
psutil = "^5.9.6"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
live-ad-web = "live_ad_detection.web_interface.app:main"
live-ad-touch = "live_ad_detection.touchscreen_ui.app:main"
live-ad-setup = "live_ad_detection.scripts.setup:main"
